+++ App/Components/ButtonDefault.spec.js
@@ -3,9 +3,9 @@ import test from 'ava';
 import React from 'react';
 import { shallow } from 'enzyme';
 import { spy } from 'sinon';
-import {FullButton} from './FullButton';
+import {ButtonDefault} from './ButtonDefault';
 
-const wrapper = shallow(<FullButton onPress={() => {}} text="hi" />);
+const wrapper = shallow(<ButtonDefault onPress={() => {}} text="hi" />);
 
 test('component structure', (t) => {
   t.is(wrapper.name(), 'TouchableHighlight');
@@ -15,7 +15,7 @@ test('component structure', (t) => {
 
 test('should call onPress', (t) => {
   const spiedOnPress = spy(() => {});
-  const wrapperPress = shallow(<FullButton onPress={spiedOnPress} text="hi" />);
+  const wrapperPress = shallow(<ButtonDefault onPress={spiedOnPress} text="hi" />);
 
   t.is(spiedOnPress.called, false);
   wrapperPress.simulate('press');
@@ -24,7 +24,7 @@ test('should call onPress', (t) => {
 
 test('should not call onPress when disabled', (t) => {
   const spiedOnPress = spy(() => {});
-  const wrapperPress = shallow(<FullButton onPress={spiedOnPress} text="hi" disabled />);
+  const wrapperPress = shallow(<ButtonDefault onPress={spiedOnPress} text="hi" disabled />);
 
   t.is(spiedOnPress.called, false);
   wrapperPress.simulate('press');
